Soft delete for entities Space, Project and Event;

+++ src/protected/application/themes/active/views/panel/part-agent.php
@@ -14,7 +14,12 @@
     <div>
         <a class="action" href="<?php echo $entity->editUrl; ?>">editar</a>
         <?php if(!$entity->isUserProfile): ?>
-            <a class="action" href="<?php echo $entity->deleteUrl; ?>">excluir</a>
+        
+            <?php if($entity->status === \MapasCulturais\Entities\Agent::STATUS_ENABLED): ?>
+                <a class="action" href="<?php echo $entity->deleteUrl; ?>">excluir</a>
+            <?php else: ?>
+                <a class="action" href="<?php echo $entity->undeleteUrl; ?>">recuperar</a>
+            <?php endif; ?>
         <?php endif; ?>
     </div>
 </article>
