+++ src/protected/application/themes/active/views/panel/part-space.php
@@ -9,10 +9,15 @@
 		<div><span class="label">Tipo:</span> <?php echo $entity->type->name?></div>
 		<div><span class="label">Área de atuação:</span> Cinema</div>
 		<div><span class="label">Local:</span> <?php echo $entity->endereco?></div>
-		<div><span class="label">Acessibilidade:</span> Sim</div>			
+		<div><span class="label">Acessibilidade:</span> Sim</div>
 	</div>
     <div>
         <a class="action" href="<?php echo $entity->editUrl; ?>">editar</a>
-        <a class="action" href="<?php echo $entity->deleteUrl; ?>">excluir</a>
+        
+        <?php if($entity->status === \MapasCulturais\Entities\Space::STATUS_ENABLED): ?>
+            <a class="action" href="<?php echo $entity->deleteUrl; ?>">excluir</a>
+        <?php else: ?>
+            <a class="action" href="<?php echo $entity->undeleteUrl; ?>">recuperar</a>
+        <?php endif; ?>
     </div>
 </article>
